Перестройка логики избранного
