--- src/components/card/index.ts
@@ -1,2 +1,0 @@
-import Card from "./Card";
-export default Card
