+++ src/components/headphones/index.ts
@@ -1,0 +1,2 @@
+import Headphones from "./Card";
+export default Headphones
