+++ src/components/headphones/info/Info.css
@@ -1,22 +1,22 @@
-.headphones__info{
+.card__info{
     display: flex;
     justify-content: space-between;
     margin-top: 55px;
 }
-.headphones__info-title{
+.card__info-title{
     font-weight: 600;
     font-size: 17px;
     color: #1c1c27;
 }
-.headphones__info-price{
+.card__info-price{
     text-align: center;
 }
-.headphones__price-after{
+.card__price-after{
     font-weight: 600;
     font-size: 17px;
     color: #ffa542;
 }
-.headphones__price-before{
+.card__price-before{
     font-weight: 600;
     font-size: 13px;
     text-decoration: line-through;
