+++ src/components/headphones/info/index.ts
