+++ src/pages/home/interfaces.ts
@@ -8,5 +8,5 @@ export interface Idata {
     name: string;
     priceAfterDiscount: string;
     priceBeforeDiscount: string;
-    isSaveLocal? : boolean
+    isFavorited : boolean
 }
