--- src/pages/home/sections/headphones/button/Button.css
@@ -1,6 +1,0 @@
-.headphones__favorite{
-    border: 0;
-    width: 20px;
-    height: 20px;
-    background-color: #fff;
-}
