--- src/pages/home/sections/headphones/button/index.ts
@@ -1,3 +1,0 @@
-import Button from "./Button";
-
-export default Button
