--- src/pages/home/sections/headphones/index.ts
@@ -1,2 +1,0 @@
-import Headphones from "./Headphones";
-export default Headphones
